Client: Remove all redundant casts
This is an automated change done by IntelliJ.

+++ client/src/main/java/agolf/lobby/LobbyPanel.java
@@ -344,7 +344,7 @@ public class LobbyPanel extends Panel {
     }
 
     protected String getTime(int var1) {
-        return var1 == 0 ? this.gameContainer.textManager.getGame("LobbyReal_TimeLimitNo") : this.gameContainer.textManager.getTime((long) var1);
+        return var1 == 0 ? this.gameContainer.textManager.getGame("LobbyReal_TimeLimitNo") : this.gameContainer.textManager.getTime(var1);
     }
 
     protected void addMessage(String var1) {
